+++ app/lib/features/home/widgets/game_card.dart
@@ -150,5 +150,4 @@ class GameCard extends StatelessWidget {
       padding: 0,
     );
   }
-
 }